+++ core-ng/src/main/java/core/framework/api/mongo/MongoCollection.java
@@ -4,6 +4,7 @@ import org.bson.conversions.Bson;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.function.Consumer;
 
 /**
  * @author neo
@@ -24,6 +25,8 @@ public interface MongoCollection<T> {
         query.filter = filter;
         return find(query);
     }
+
+    void forEach(Query query, Consumer<T> consumer);    // mongo driver fetches results in batch
 
     <V> List<V> aggregate(Class<V> resultClass, Bson... pipeline);